Add `GitHubWebhookTask`
