+++ plugins/github/server/index.ts
@@ -5,6 +5,7 @@ import { GitHubIssueProvider } from "./GitHubIssueProvider";
 import router from "./api/github";
 import env from "./env";
 import { GitHub } from "./github";
+import GitHubWebhookTask from "./tasks/GitHubWebhookTask";
 import { uninstall } from "./uninstall";
 
 const enabled =
@@ -20,6 +21,10 @@ if (enabled) {
       ...config,
       type: Hook.API,
       value: router,
+    },
+    {
+      type: Hook.Task,
+      value: GitHubWebhookTask,
     },
     {
       type: Hook.IssueProvider,
